ModelDetails panel now listens for Model updates from the store
Most of the basic fields in the ModelDetails panel are rendered from this.props.entity. Formsy considers the "pristine" value of each field to be the "value" prop when it was rendered. The prop contents is fed from this.props.entity in ModelDetails, which is tied to the stored instance of the model. So, in effect, after the form is submitted, the new values of simple fields automatically become the new pristine values.

However, the model properties and relations are added dynamically and are not rendered directly from props; they are rendered from this.state in ModelDetails. The problem is that this.state was only set once in the constructor. Adding/removing properties/relations did modify this.state, but if a property or relation *field* was changed, that change never made it back into this.state in ModelDetails. As a result, the Formsy inputs were never re-rendered with a new pristine values.

At least that what I *think* was happening.