Ensuring that the ID is always passed to the constructor of BaseModel
All BaseModel subclasses have constructors, but they are never used. All instantiation happens via `BaseModel.create()`, which just calls `new this()`. The problem is that this ends up creating a temporary ID, which then potentially gets overridden when properties are copied into the new object by `create()`. Any code in the constructor that uses `this.id` will be using a temporary ID that disappears immediately the constructor runs.

This was causing problems for `Pipeline` constructor, which creates Port objects in the constructor and passes `this.id` to them.